requirements.txt
numpy==1.9.2
cassandra-driver==2.7.2
mysql-connector
rstr==2.2.3